XMLHttpRequest/tests/submissions/Opera/abort-during-done.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: abort() during DONE</title>
    <script src="/resources/testharness.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest(),
            control_flag = false,
            result = [],
            expected = [1, 4] // open() -> 1, send() -> 4
        client.onreadystatechange = function() {
          test.step(function() {
            result.push(client.readyState)
          })
        }
        client.open("GET", "resources/delay.php?ms=200", false)
        client.send(null)
        assert_equals(client.readyState, 4)
        client.abort()
        assert_equals(client.readyState, 0)
        assert_array_equals(result, expected)
        test.done()
      })
    </script>
  </body>
</html>
